Backend/grocapitalapp-django/requirements.txt
asgiref==3.3.1
astroid==2.5.1
autopep8==1.5.5
backports-datetime-fromisoformat==1.0.0
backports.zoneinfo==0.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
Django==3.1.7
django-cors-headers==3.7.0
django-environ==0.4.5
django-rest-knox==4.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
ecdsa==0.14.1
enum-compat==0.0.3
freshbooks-sdk==0.6.1
future==0.18.2
idna==2.10
intuit-oauth==1.2.4
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
oauthlib==3.1.0
Pillow==8.3.1
plaid-python==7.4.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.7.2
python-jose==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
wrapt==1.12.1
xmltodict==0.12.0
